chore(cli): dbAuthHandler() (ESM prep)
